- fix bug where value / number is same as first

+++ src/index.ts
@@ -25,17 +25,20 @@ form.addEventListener('change', (e: Event) => {
 
 dotBtn.addEventListener("click", () => {
   if(screen.value == "Syntax Error") return
+  if(screen.value == "NaN") return
   MyCalculator.addDot().inputFocusToEnd()
 })
 
 calculateBtn.addEventListener("click", () => {
   if(screen.value == "Syntax Error") return
+  if(screen.value == "NaN") return
   if(screen.value === "") return
   MyCalculator.calculate()
 })
 
 deleteBtn.addEventListener("click", () => {
   if(screen.value == "Syntax Error") return
+  if(screen.value == "NaN") return
   MyCalculator.delete()
 })
 
@@ -46,6 +49,7 @@ resetBtn.addEventListener("click", () => {
 normalBtns.forEach((normalBtn: HTMLButtonElement) => {
   normalBtn.addEventListener("click", () => {
     if(screen.value == "Syntax Error") return
+    if(screen.value == "NaN") return
     MyCalculator.disPlayButtonToScreen(normalBtn.value).inputFocusToEnd()
   })
 })
